move classes into subfolder fix some problems and mark TODO
