+++ library/Director/Backend/MonitorBackend.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Icinga\Module\Director;
+namespace Icinga\Module\Director\Backend;
+
+use Icinga\Data\Filter\Filter;
 
 interface MonitorBackend
 {
@@ -8,7 +10,11 @@ interface MonitorBackend
 
     public function hasHost($hostname);
 
+    public function hasHostWithExtraFilter($hostname, Filter $filter);
+
     public function hasService($hostname, $service);
+
+    public function hasServiceWithExtraFilter($hostname, $service, Filter $filter);
 
     public function getHostLink($title, $hostname, array $attributes = null);
